requirements.txt
webrtcvad-wheels
fastapi>=0.110
uvicorn[standard]>=0.27
sqlalchemy>=2.0
pymysql>=1.1
pydantic>=2.5
pydantic-settings>=2.2

# Auth
PyJWT>=2.8

# LLM (DeepSeek OpenAI-compatible SDK)
openai>=1.30

# Speech (iFlytek WebSocket)
websocket-client>=1.7
certifi>=2024.2

# WS demo client
websockets>=12.0

# S3 storage
boto3>=1.34

# DB migrations
alembic>=1.13

# Optional: better VAD (install only if needed; otherwise fallback to EnergyVad)
# webrtcvad>=2.0.10

# MQTT
paho-mqtt==1.6.1
